Added next version of foofen
I made a lot of progress with this one. I added some UI that appears at the end of an NPC dialogue that changes changes scenes. I also added a script that follows the player so now there's more room to work with, I even added two UI images to help the user experience that prompts the user to press E or R